+++ R/statistics.R
@@ -126,6 +126,27 @@ gmean <- function(x, trim = 0, na.rm = FALSE) {
   exp(mean(log(unclass(x)[index]), trim = trim, na.rm = na.rm))
 }
 
+# Quantile =====================================================================
+#' @export
+#' @method quantile CompositionMatrix
+quantile.CompositionMatrix <- function(x, ..., probs = seq(0, 1, 0.25),
+                                       na.rm = FALSE, names = TRUE) {
+  apply(
+    X = x,
+    MARGIN = 2,
+    FUN = stats::quantile,
+    probs = probs,
+    na.rm = na.rm,
+    names = names,
+    ...
+  )
+}
+
+#' @export
+#' @rdname quantile
+#' @aliases quantile,CompositionMatrix-method
+setMethod("quantile", "CompositionMatrix", quantile.CompositionMatrix)
+
 # Scale ========================================================================
 #' @export
 #' @method scale CompositionMatrix