+++ NEWS.md
@@ -1,7 +1,8 @@
 # nexus 0.1.0.9000
 ## New classes and methods
-* Add `split()` methods to divide into groups.
 * Add `hist()` methods to plot ILR histograms.
+* Add `quantile()` methods for compositional data matrix.
+* Add `split()` methods to divide into groups.
 
 ## Enhancements
 * `transform_ilr()` gained a new argument allowing the specify the base of the transformation.